Fix overlapping ranges.
Sometimes we add a range that overlaps with two or more existing ranges.
The old way only modified the first range, but with this change we
check the next ranges as well.

I don't know that this bug caused any problems, but it made the debug
output look a bit odd sometimes.

Signed-off-by: Dan Carpenter <error27@gmail.com>