introduce smatch_containers.c
Some false positives are caused code like this:
kfree(foo->bar);
kfree(foo);
foo = something();
kfree(foo->bar);

It's not a double free because the container of foo-bar was
redefined.  Smatch containers lets check scripts revert to a default
state when the container is redefined.  This only works with
set_state_expr() though.  For the existing checks the only places
which should be using set_state() are for things which don't have
containers anyway.

The key new function for checks to put in there setup is this.
void set_default_state(int owner, struct smatch_state *state);

Signed-off-by: Dan Carpenter <error27@gmail.com>