Add check_assigned_expr.c
This isn't a check by itself, it's just a tool.

When I was looking for buffer overflows I wondered what allocation
functions people were using to allocate pointers used as arrays.
Code like this:
int *foo;
foo = some_allocation_func();
foo[x] = 1;
I wrote a script to print out the "= some_allocation_func()" after it
saw "foo[x]".  This script is useful for that.

I thought about disabling it by default but decided to enable it, so
it's easy to use later.

Signed-off-by: Dan Carpenter <error27@gmail.com>