Improve check_err_ptr.c
check_err_ptr.c prints a list of functions that return an ERR_PTR().

The old code assumed that if you called ERR_PTR() you returned an
ERR_PTR() but that's not always true.  The new code only prints an
error if you actually return an ERR_PTR().

The disadvantage is that it might miss some.  I think when I improve
the loop handling it will catch basically everything though...

Signed-off-by: Dan Carpenter <error27@gmail.com>