Dual pass system [patch #1]
Smatch has always had an issue with loops and this patch starts to
address that.  To handle loops we need to go through the code twice
and that's what this patch does.  Going through twice helps with
goto statements but not much else.  The next patches will add
handling for loops.

Signed-off-by: Dan Carpenter <error27@gmail.com>