Optimization: prepend() -> append()
prepend() used to move all the elements in the array one by one to
create a blank space at the front.  append() uses strlen() and strcat().

It seems like there should be a way to do the strlen and strcat in
one step instead of two separate functions but I don't know what it is.
And no one on #offtopic2 responded to my question.  Screw you guys and
your whole American timezone.

Signed-off-by: Dan Carpenter <error27@gmail.com>