Modify the return hook to pass an expression instead of the statement.
It was silly to pass the whole statement instead of just the ret_value
expression.

The notable thing about this patch is that it is now possible to pass
a NULL value to clients.  (ret_value is sometimes NULL).  Now every hook
function needs to assume that it is possible to receive a NULL parameter.

Signed-off-by: Dan Carpenter <error27@gmail.com>