try handle do {} while (i-- > 0)
So the problem is that by default we are still going only one time
through loops.  So if we have code like:
x = 10;
do { } while (--x >= 0)
When it gets to the comparison smatch says that x can only possibly be
9.  The  other path is impossible.

Now we say, Ok.  It's going to be any number less than zero.

Signed-off-by: Dan Carpenter <error27@gmail.com>