wine: check for locking bugs
Quite a bit of this code came from the check for kernel locking functions.
The good thing about this code is that you can check for more locks by
adding a couple lines to the lock_table.  It also will print an error if
you use the wrong unlock function so if you lock with
RtlEnterCriticalSection(), you can't unlock with LeaveCriticalSection().

So it's improved from the kernel locking check and eventually that should
be merged into this check.

Signed-off-by: Dan Carpenter <error27@gmail.com>