check_deref_check: get rid of some false positives
In the new version it's always OK to dereference the variable when we
know it's not null.

Also we don't print an error message if we know the variable is non null
when we check.  Mostly it's macros are the ones checking here.  If not
then it's probably just needless checking rather than actual null
dereference bugs.

Signed-off-by: Dan Carpenter <error27@gmail.com>