clean up kfree_skb() check.
Deleted all the old functions to figure out types.  They were messy and
not as good as the new get_type().

I also modified the test to use get_variable_from_expr_complex() so it
will catch more errors potentially.  And even if that function returns
null but we know the type we can still print an error.

Improved the validation script by adding the xx.skb.

Signed-off-by: Dan Carpenter <error27@gmail.com>