fixes for unified wine & kernel locking script
Handle locks where the RETURN_VAL is locked.  Before we weren't using the
function_assign_hook() so it was totally wrong.

Added a bunch of lock functions with a double underscore on the front and
shifted things around a bit.

Also added a bunch more entries to the table for handling irq disable
and enable.

One thing which is interesting is that depending on the configuration
_spin_lock_irqsave can take either 1 or 2 options.
+ 	{"_spin_lock_irqsave", LOCK, "irqsave", RETURN_VAL, ret_any},
+ 	{"_spin_lock_irqsave", LOCK, "irqsave", 1, ret_any},

There are probably still quite a few bugs left.

Signed-off-by: Dan Carpenter <error27@gmail.com>