check_overflow:  separate the two types of states
It turns out that there were two kinds of states we were tracking.  The
size of the array and the array offset max.

I don't think this would have caused any bugs because the array is not
going to be used as an offset.  But anyway, it's easier to understand
this way.

Signed-off-by: Dan Carpenter <error27@gmail.com>