smatch_extra: I don't like empty ranges
Originally I thought empty ranges should mean everything was included.  But
these days I think that was a mistake.  Most empty ranges come from
impossible conditions.

x = 0;
if (x > 100) {
    // <-- x has an empty range.

Then when we merge the ranges we get an empty range again and we can't
get rid of the empty range because it trumps everything.

Really --two-passes and --known-conditions should be turned on by default
and smatch should complain about impossible conditions that can not be
optimized away at compile time.

Signed-off-by: Dan Carpenter <error27@gmail.com>