Fix bug in handling for loops
This assuming you haven't passed --two-passes.

Say we have a pre loops like: for (i = 0; i < 10 && something(); i++) { ...
If we didn't have the "something()" condition then smatch would handle it
as a special case and it would actually get this right.

The correct thing is to parse the condition before the loop and save all the
false states, then to parse the condition after the loop and save the states.
The false states from both passes get merged together and that is the
states after the loop.

The original code parsed the condition only once.  Then it did a silly
hack by using __use_false_only_stack() to combine the states that had
changed inside the loop with the states which were false at beginning.

This was the only code that called __use_false_only_stack so I will
delete that code in the next commit.

Signed-off-by: Dan Carpenter <error27@gmail.com>