add --rare option for rare bugs
Adding checks to smatch slows it down, so sometimes it's not worth doing
the check every time.  For example, I wrote a check for reusing skb after
calling dev_queue_xmit(skb).  This was based on a patch in 2.6.33-rc4.
When I tested the check on 2.6.33-rc3 it found the bug which the patch
fixed but it didn't find any others.

Whenever there is an -rc1 release I will pass the --rare option to look for
bugs like that.

Signed-off-by: Dan Carpenter <error27@gmail.com>