function_implies_hook: fix bug when a function has 2 implications
Implications are formed when we merge slists.

It's like standing in the middle of a river looking upstream.  If we know
that we dumped in blue die and red die together in one tributary and we
see a part of the stream that has blue and red die we know where that water
came from.

For implication hooks the true path is one tributary and the false path is
another.

With the old code there were more than one possible way for the tributaries
to flow into the main river.  Really we just wanted one.

Signed-off-by: Dan Carpenter <error27@gmail.com>