check_signed:  clean ups.  add check for impossibly low comparisons
There was already a check for comparing unsigned numbers to negative.  I
left that one the same.  Also now if you compare a char to -1000 or other
impossibly low numbers it will complain.

The bit that printed out "this condition is always true/the same/false was
a bit buggy so I fixed it and cleaned it up a bit.

Signed-off-by: Dan Carpenter <error27@gmail.com>