cleanup:  get rid of smatch_containers.c
smatch_containers.c is dead code.

smatch_containers was a good idea in general but the implementation was
rubbish and so it was never used.

Really it's just a subset of smatch_modification_hooks.c which are used.
smatch_modification hooks are more flexible because you pass a function
instead of a state.  One way this flexibility is used in the current code
is that it lets you delete the state if you want.

So when we re-implement smatch_containers it will be different enough that
keeping the current code around doesn't help.

Signed-off-by: Dan Carpenter <error27@gmail.com>