introduce: set_default_modification_hook()
A lot of the time you end up calling add_modification_hook() for every
single set_state() or set_state_expr().  This is a helper function that
takes care of it for you.

What surprised me is that quite in quite a few places the code didn't
call add_modification_hook() for every set state.  I only modified the
checks where every set_state() was used with an add_modification_hook()
so this change shouldn't make any difference to runtime.

Signed-off-by: Dan Carpenter <error27@gmail.com>