partly replace check_null_deref.c
I really hate check_null_deref.c.  After this change it will only be turned
on if you use the --spammy option.  The problem is that it tries to do too
much.  It marks things as null if they are uninitialized, or set to null,
or certain unmatched states get marked as null, or some function parameters
or sometimes if we test a variable to see if it is null.  It's a mess.

I've written a new check called check_deref.c, but it only marks variables
as null if they are uninitialized, set to null or assigned a potentially
null return value.  The other things are important and I probably will
create a check for them, but it will be separate checks and I won't try
cram everything together like before.

The new test has much more straight forward logic.  The code is cleaner.
The error messages are better.

Signed-off-by: Dan Carpenter <error27@gmail.com>