tweak unreachable code messages
The problem is that smatch was printing unreachable code for declarations
at the start of switch blocks before the first case label.  This fixes that.

That message was originally written as debug information to let people know
that set_state() was being ignored.  It has found some bugs so I left it
visible when I hid most of the other debug information.

Still maybe the test in smatch_flow.c which prints about unreachable code
is better.  It was disabled before because it printed too much stuff, but
I have enabled it again now when you pass --spammy.

Signed-off-by: Dan Carpenter <error27@gmail.com>