bail if things are taking too long
I was trying to test postgresql and there is a bit of code that does:
    if (!strcmp(foo, bar) || !strcmp(x, y) ...
It does around a million strcmp comparisons in a row.  And glibc strcmp()
is a very complicated macro.  Smatch got bogged down trying to figure out
all the various code paths and macros.

This patch says just bail on that function and go to the next.

Signed-off-by: Dan Carpenter <error27@gmail.com>