add smatch_scripts/new_bugs.sh: prints the bugs introduced between 2 smatch runs
Say you want to see the bugs introduced between the -rc1 and -rc2 kernel,
you could run smatch on each kernel.  Then use new_bugs.sh to show the
newly introduced bugs.

The advantage of the script over a normal diff is that it doesn't print
out bugs where only the line number changed but the message is the same.

Signed-off-by: Dan Carpenter <error27@gmail.com>