smatch_flow: change how the big expression stack works a bit.
Before nothing was removed from the stack.  This was because of laziness.
It is fixed now.

Also handling loops is another way of splitting an expression.  The
pre loop and post loop conditions didn't get added to the big expression
stack.  I added them now in the smatch_conditions.c file.  It seemed easier
that way, but I may change this in the future.

Signed-off-by: Dan Carpenter <error27@gmail.com>