check_overflow:  don't complain about if (ptr < &array[ARRAY_SIZE(array)])
This is a common way of looping over an array.

	for (p = a; p < &a[ARRAY_SIZE(a)]; p++)
		printf("%d\n", *p);

I was strict about which places to not complain about.  The offset has
to be known at compile time.  It has to be a comparison.

The bad bit is that if you dereference p after the end of the array
smatch will not complain.  That should maybe be added to another check.

Signed-off-by: Dan Carpenter <error27@gmail.com>