extra: comparing something to zero doesn't mean it can be zero
Before if you had code like:
int x = 42;

if (x == 0)
	printk("error\n");

Smatch would think that it was possible for x to be equal to zero at the
end.  Really it could just be a paranoid macro.

It might be possible if it were in a loop actually.  But for now let's go
with the one that cuts down on false positives.

Signed-off-by: Dan Carpenter <error27@gmail.com>