handle for loops where we don't know the upper limit
Smatch handling of loops has always been problematic.

Before this change if we had a loop like this:
for(i = 0; i < x; i++)
        __smatch_value("i");

It would print "i == 0".  After the change it will print 0 to x if x is
known or 0 to max.

This isn't a complete solution, but it's an improvement.

Signed-off-by: Dan Carpenter <error27@gmail.com>