equiv, implied: fix handling borrowed implications
With the equiv code if "x" implies that "y" is 1 and we say:
	foo = x;
then that means that "foo" implies "y" is 1.

We do this by copying the ->left and ->right parts of the sm_state.

Implications work by merging states together from the my_pools which are
true.

The problem is that we borrowed the ->left and the ->right.  Those have the
name set to "x".  So we end up with a merged state for "x" instead of "foo".
So force it back to "foo" at the end.

Also now that we change the name, we can't count on finding our state in
->my_pool.

Signed-off-by: Dan Carpenter <error27@gmail.com>