extra: assume impossible things are impossible
I've waffled back and forth on how to handle this:

	if (x < 0) {
		if (x == 10)
			__smatch_value("x");
	}

What should smatch print there?  One approach has been to say that maybe
smatch was wrong before and x can be 10.  But my new approach which I've
been moving to is to say, "It's probably just a macro asking if x is 10.
let's assume that the check is bogus.  x = (empty)"

Signed-off-by: Dan Carpenter <error27@gmail.com>