locking: add more standard locks
For some kernel releases spin_lock() was always a macro.  Now it's a
function.

I decided that all of the functions that start with a single underscore
probably have a no underscore version as well.  They could be macros now
but they may become real functions later on.  So I added them all.

Signed-off-by: Dan Carpenter <error27@gmail.com>