equiv: eliminate bogus call to set_state()
remove_from_equiv() is called when two variables are no longer equivalent.
The caller is supposed to update the state for the variable that was
modified, we are not supposed to do it here.  In fact we can't do it here
correctly.

This patch shouldn't really make a difference from what I can see.  The
original code set the state to a bogus value and then immediately overwrote
it with the new value.

Signed-off-by: Dan Carpenter <error27@gmail.com>