*new* freeing_null: calling free for a variable that's always NULL
This is only turned on with --spammy because it has quite a few false
positives.  They are caused by loops and because smatch doesn't do
cross function analysis.

It finds quite a few places that have freeing NULL variables as part of
their error handling.  Also it finds a couple bugs where the wrong variable
is freed.

Signed-off-by: Dan Carpenter <error27@gmail.com>