bail out earlier if functions get too difficult
This only bails out on one function, and it still doesn't make my kernel
compile on my Eee PC work.  But it's an improvement.

Mostly we bail on functions that have huge case statements.

It's weird because the old higher limits used to bail on a lot of functions
in my allmodconfig but now they don't bail on any.  Perhaps someone changed
the code?  I'd feel bad if  people were changing code just so smatch
worked on my Eee.

This stuff needs to be fixed properly at some point.

Signed-off-by: Dan Carpenter <error27@gmail.com>