Allow parsing L'\0'
It is nasty that L'\0' start from an identifier letter. I try my best
not to slow down the hot path. The test is done inside get_one_identifier()
after the ident hash is built.  It look a little bit out of palace but
faster than testing 'L' before hand.

Signed-off-by: Christopher Li <sparse@chrisli.org>