*new* check_kernel.c: handle tomoyo_memory_ok() and friends
If tomoyo_memory_ok(xxx) returns true, that means that xxx is a non null
pointer.

I created check_kernel file to update SMATCH_EXTRA with this sort of
information.  I used the return_implies_state() hooks to do this.  When we
update "xxx" to say that it is non null then we have to update the
equivalent states as well so I created set_extra_expr_nomod().

IS_ERR_OR_NULL(xxx) is a similar function but when it returns true we
don't know that xxx is NULL, so we don't set that side.  This generates a
warning in smatch_function_hooks.c so I removed the warning code.

Also I updated check_deref.c to not print errors if smatch_extra said a
pointer is non-null.

Signed-off-by: Dan Carpenter <error27@gmail.com>