extra: handle i++ and i += 2 better outside loops
This patch adds the function inside_loop() which returns 1 if we are
inside a loop or if we have reached a goto label.

Inside loops we assume that i++ can go up to max (Unless they are in a
canonical loop format, that's handled as a special case).  Outside of a
loop i++ is just a matter of adding 1 to the min and max values of i.

At this point "i = i + 1;" is always treated as being outside a loop
because smatch isn't clever enough to see that it's adding itself to
itself.

Signed-off-by: Dan Carpenter <error27@gmail.com>