*new* check_no_effect: print warnings statements that have no effect
I saw someone make a bug like this on lkml, but it was corrected
immediately by a reviewer.  I think there might be a version of gcc that
warns about these.

This worked on my test files but when I tested the kernel I didn't find
any bugs, or any false positives.

Signed-off-by: Dan Carpenter <error27@gmail.com>