flow: the next statement should never be the current statement
The next statement stuff is imprecise.  That's ok because we only care
about the location of it in a file, so we can tell when we're inside a
macro.

The problem here is that if we're in a block with one statement the next
is the same as the current and the in_macro() function returns true when
it shouldn't.

Signed-off-by: Dan Carpenter <error27@gmail.com>