*new* smatch_scripts/unlocked_paths.pl
I was looking for race conditions caused by not taking a lock so I wrote
this script.

Say if you are always supposed to call frob() with with @foo_lock held then
first generate a warns.txt using the --call-tree option.  Then the command
is: "unlocked_paths.pl warns.txt foo_lock frob"

It doesn't handle calling function pointers.  The base of the path will
probably be something that gets assigned to a function pointer actually.

Signed-off-by: Dan Carpenter <error27@gmail.com>