precedence: don't complain about  if foo is 0-1
Previously smatch would not complain about "if (!foo = bar()) {" if
either "foo" or "bar()" was of type boolean.  But sometimes people use
variables as boolean even though the type is not declared as boolean.

If we can tell that foo is only either 0 or 1 even without looking at the
type then do not complain.

Signed-off-by: Dan Carpenter <error27@gmail.com>