ranges: introduce possibly_true/false_range_lists()
We used to have possibly_true_range_list() but I renamed that to
possibly_true_range_list_lr().  This is a temporary thing.  I plan to
remove it in later commits.

The new functions are basically the same as the old
possibly_true_range_list() except instead of a dinfo they take a range.
I plan to use them more extensively in later commits.

Signed-off-by: Dan Carpenter <error27@gmail.com>