implied: allow the known part to be a range list
So implications basically works like this:

if (x == 42) { ...

If x is a merged state then it potentially has implications.  So we take
the 42 and the pools from x and figure out where the condition is true.

But what about if, instead of 42 we have y?  In the current code if y is
a single value like this:

y = 42;
if (x == 42) { ...

The current code can handle that.  The thing it can't handle is if y is
a range like 4-10.

This patch tries to address that.

Signed-off-by: Dan Carpenter <error27@gmail.com>