macro_table.c: replacement for in_macro()
in_macro() was an ugly hack, it worked OK for the kernel but not for user
space glibc.  The new code hooks into the pre-processor and is more
accurate, you can also find out the name of the macro which could be
useful.

Instead of in_macro() use get_macro_name().  It returns null if we're not
inside a macro.

Signed-off-by: Dan Carpenter <error27@gmail.com>