flow: handle logic things like: foo(a && a->bar);
Smatch already handled select statements like:

        foo(a ? a->bar : 0);

but it didn't handle logicals or comparisons until now.  To bad news is
that handling these things causes some false positives in the broadcom
driver because they use functions instead of macros to do asserts.
Fixing that is quite a bit of work.  The good news is that this change
fixes some things when using smatch to test Wine.

Signed-off-by: Dan Carpenter <error27@gmail.com>