extra: tweak assignment of unsigned values
Smatch extra tends to save everything as potentially having min-max values,
where min and max are the min and max for long long type.  I would like to
move away from that, but it requires some thought on how to do it without
breaking get_fuzzy_max().

Anyway, this change just means that if you have an unsigned on the right
then the left is unsigned.  It's not technically correct because if you
have a big unsigned on the right that can be a signed value...  But it
doesn't seem to cause any problems so I'll merge it for now.

This all requires some more thought.

Signed-off-by: Dan Carpenter <error27@gmail.com>