overflow: ignore arrays with only one element.
It's pretty common to put a one element array at the end of an struct:

	struct foo {
		int size;
		char buf[1];
	};

The old check had logic to handle this but it failed because xfs put their
resizable buffer in the middle of the array.  XFS scares me and I didn't
look into how this works.  Also eicon initializes all their strings to
point to a null character and that confuses smatch.  So now I've removed
the struct logic and just ignore all one element arrays.

Signed-off-by: Dan Carpenter <error27@gmail.com>