snprintf: turn off be default
Turn the snprintf check off unless --spammy is used.

This check is meant to find places which don't take into account that the
return value of snprintf() can be larger than the buffer size.  It works
as advertised but it turns out that it never finds any actual bugs.  It
does find some crap code where the author could just as well have used
sprintf() but it doesn't find actual overflows.

Signed-off-by: Dan Carpenter <error27@gmail.com>