extra: ignore WARN_ON()
People put warn on macros into code and it means that they don't expect
that condition to be true.  Smatch tends to see it the opposite way that
you are testing for it and so it can be true.

Smatch_extra is the code that tracks what possible values a variable can
be.  With this change, smatch_extra ignores everything done inside a
WARN_ON() macro.  This change mostly affects the buffer overflow check.

        WARN_ON(foo >= ARRAY_SIZE(array);
        array[foo] = 42;

The old code would print a warning and the new code doesn't.

Signed-off-by: Dan Carpenter <error27@gmail.com>