overflow: handle "if (strlen(foo) < 10) {..."
We're only storing the max size of the string.  Smatch extra stores ranges
but in this case we're using alloc_my_state() and we're only storing
the max.  And unfortunately only one side is maxed out.

if (strlen(foo) < 10) {
    //  foo is max of 10 chars (including the NULL)
} else {
    // there is no max here
}

It would be nice to store a range so we could do stuff like get_fuzzy_max()
but oh well...  Good enouch for now.  And most code only cares about the
max so it could be worse.

Signed-off-by: Dan Carpenter <error27@gmail.com>