implied: only print the nr_children messages if there is a warning later
The nr_children messages basically mean that Smatch is having a hard time
figuring out implications.  It's normally caused by huge switch statements
or really complicated macros.

The messages are only helpful if Smatch prints a warning.  (It tells us
that it's possibly a false positive resulting from a break down in
implications).

Signed-off-by: Dan Carpenter <error27@gmail.com>