*new* uncapped_alloc: print places where the user decides how much memory to alloc
This test isn't the greatest...  Basically the idea here was to do an
audit for places that don't limit the amount of memory that a user can
allocate.

I was looking for bugs like: 5b75c4973c "can: add limit for nframes and
clean up signed/unsigned variables".  This check does find that bug and it
finds one other bug.

It turns out that normally if you let the user decide how much memory to
allocate the worst thing which can happen is that the kernel prints a
stack dump or you might get a DoS.

Signed-off-by: Dan Carpenter <error27@gmail.com>