buf_size: do cross function analysis
Use smatch_db to find the largest buffer that is passed to a function
parameter and store that as the buffer size.

Ideally we would have a column for the call expression pointer and we would
create an slist with the settings for each call and then merge them
together to get the actuall sizes at the start of the function.   Then
smatch would use smatch_implied to tell what the actual buffer sizes were
when we referenced the buffer.

This did catch some bugs, but it's not ideal.

Signed-off-by: Dan Carpenter <error27@gmail.com>