*new* get_user_overflow: check for integer overflows from get_user()
This looks for bugs like:

	get_user(x, user_ptr);
	if (x < 100)
		return;
	foo = x * 2;

The problem is that if we set x to a negative value we can make foo be
whatever we want, by taking advantage of the integer overflow.

There are some false positives here.  If you say:

	if (x < (unsigned int)100)
		return foo;

Then that means if x is negative it will return.  I need to fix
smatch_extra to make this work.

Also there is one place that does:

	char c;

	get_user(c, user_ptr);
	if ((int)c * 2) { ...

The cast to an int makes smatch think that c goes from INT_MIN to INT_MAX
when it's just a char.  Fixing that requires a change in smatch_extra as
well.

Signed-off-by: Dan Carpenter <error27@gmail.com>