slist: preserve the ->line value correctly
Smatch hasn't really used the ->line value up to now.  Partly that's
because it hasn't been showing the correct line number.

In add_possible() the alloc_state_no_name() is basically trying to do a
small clone of the sm_state so the line should be preserved.

In merge_sm_states() it doesn't make sense to try preserve the ->line
value.  What we want is the the line where the merge happens.

Signed-off-by: Dan Carpenter <error27@gmail.com>