smatch: add --data=<dir> option
I'd like to integrate smatch scan to the my project. Unfortunately, the
current code does not allow to specify directory for the smatch-data and
the data directory is expected outside the project source tree.

This patch add --data=<dir> to specify the path, so then it's
possible to use for example:

   checksmatch:
   	make CHECK="smatch --data=$(top_srcdir)/anydir" CC=cgcc

in project Makefiles and then use "make checksmatch".

Signed-off-by: Karel Zak <kzak@redhat.com>