flow: pass condition assignments to clients with ASSIGNMENT_HOOK
The affects code like this:
	foo = !bar;
where "foo" is either one or zero depending on the value of "bar".

In __handle_condition_assigns() it sets the value of "foo" for SMATCH_EXTRA,
but there are other checks need to handle the assignment as well.  So I
added a __pass_to_client(expr, ASSIGNMENT_HOOK) and I added a check in
the smatch_extra.c hook so it would ignore the assignment.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>