no_effect: add some more tests
This test is sort of pointless since it duplicates functionality from gcc.
But I already wrote it, so I'm going to commit it.

What happened is that someone submitted a patch with a bug like this so I
assumed that gcc didn't catch it.  But then I didn't find any similar bugs
in the kernel, and I investigated some more and it turns that that the
submitter just ignored the warning or was using an old version of gcc or
something.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>