*new* min_t() casting that truncates the values
Say you have code like this:
	unsigned int x, y;

	y = min_t(int, x, 100);

The intent was that y would be a number between 0 and 100.  But the way
min_t() works is that it casts x to int first so large values become
negative and thus less than 100.  The result is that y can be a very large
value.

This test produces a lot of false positives where the value of x cannot
be larger than INT_MAX because it's checked outside the function.  For
example, the amount of data a user wants to read from a file is given
in size_t (unsigned long) but it's capped in the VFS at slightly under
INT_MAX.

But I feel that even though those aren't security bugs, they're still
ugly and should be corrected.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>